Initialize BlockScript to working template
[Context]
For some reason, not building the base template
is causing Log4J issues when attempting to runClient.

[Changes]
- Initialize repo with a definitely working template
  before making any changes

[Note]
If current hypothesis is correct, starting from blank
slate after changes are made may result in same Log4J
issue. This is on top of a custom JRE to correct the
M1 mac java mismatch issue.